// src/pages/first-launch/DarkDropdown.module.less.d.ts
declare const styles: {
  ddRoot: string;
  ddTrigger: string;
  ddTriggerText: string;
  ddTriggerPlaceholder: string;
  ddChevron: string;
  ddOpen: string;
  ddDisabled: string;
  ddMenu: string;
  ddMenuUp: string;
  ddOption: string;
  ddOptionFocused: string;
  ddOptionSelected: string;
  ddCheck: string;
  readonly [key: string]: string;
};
export default styles;